Trying alternative pkg show:
This would be an alternative where we are able to Show GUIX and NIX packages installed as a separate value. This is solely due to the fact that GUIX and NIX can be installed in addition to a distributions normal package manager, and it's tallies would be separate from the default package manager. Likewise if a user doesn't have them installed, they simply don't show.