Initial work on Mac Compatibility:
Added initial work on Mac Compatibility. Everything will become a case statement in this test. MacFetch, will be identical to Fetch.sh for now with the exception of new variables. This is to allow development on standard Fetch.sh while MACFetch reaches feature parity. Once that has happened. We'll re-combine them all to Fetch.sh. In addition all the Components will be CASED for MacOS.