fetch.sh Mac Compatible build 1.0:
This is in the roughest way now mac compatible. Each of the component shell scripts have been cased to handle MacOS/Darwin and other UNIX OS's in the future.

There is only the addition of one Global Variable since its used twice in both os_arch.sh and de-wm_theme.sh. It should fail gracefully on Linux.

The script won't be as fast on MacOS thanks to some of the methods I'm using to call up the information, but it is functional, we can improve from here. This was all tested on MacOS Catalina 10.15.1