Packaging Sysfetch: Many Additions:
This pull has had a very large: find and replace done upon it for the new name change, in addition it has added quite a few new things.

It adds the first instructions on how to build an installable .Deb package for distribution, a manpage for that package, as well as a contributing.md and license.md in addition to a HowToBuild.md which covers how to use fpm to package sysfetch into a .deb. I have tested the resulting deb from the fpm and the command I've build, and it installed and removed perfectly fine on Ubuntu 21.10. It's pretty neat to be able to just call it up, since it's officially a program now.

The .deb is also really tiny, mine only came out to 17.9 KB.

This request includes all the MacOS work from the last two days, as well so if that is merged into master first, this won't be behind.