small issues in README:
fixed small issues in README.md and Contributing, that weren't working with Github Markdown. or just eliminating extra .sh's from the old name.

README.md will still need to be altered when moving to master to remove some of the alpha verbiange, but we're cutting down the bulk of the changes here.